requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
